wip(ui-design-tool): move logic to InteractionController from TransformOverlay
