+++ packages/ui-design-tool/components/InteractionController/InteractionController.css.ts
@@ -1,8 +1,13 @@
-import { style } from '@vanilla-extract/css';
+import { UIInteractionElementDataAttributeName } from '@/types/Identifier';
+import { globalStyle, style } from '@vanilla-extract/css';
 
 export const root = style({
   position: 'relative',
   width: '0',
   height: '0',
   pointerEvents: 'none',
+});
+
+globalStyle(`${root} [${UIInteractionElementDataAttributeName.handleType}]`, {
+  pointerEvents: 'auto',
 });
