+++ packages/ui-design-tool/components/InteractionController/InteractionController.tsx
@@ -12,6 +12,7 @@ import {
   UIInteractionElementDataAttributeName,
   UIInteractionElementDataset,
   UIDesignToolElementDataAttributeName,
+  UIRecordKey,
 } from '@/types/Identifier';
 import { isUIRecordKey } from '@/utils/model';
 import { useEvent, useEventListener } from '@pigyuma/react-utils';
@@ -22,6 +23,8 @@ import { PointerEventsPreventer } from '../PointerEventsPreventer/PointerEventsP
 import { TransformOverlay } from '../TransformOverlay/TransformOverlay';
 import * as styles from './InteractionController.css';
 import { InteractionControllerProps } from './types';
+import useResizeFunctions from './useResizeFunctions';
+import useRotateFunctions from './useRotateFunctions';
 
 /** @todo 설계가 일정 수준 이상 확정되면: 테스트 코드 작성 */
 export const InteractionController: React.FC<InteractionControllerProps> = React.memo(() => {
@@ -29,11 +32,15 @@ export const InteractionController: React.FC<InteractionControllerProps> = React
   const uiSelectorAPI = useUISelector();
 
   const status = useStatus();
-  const hovered = useHovered();
-  const selected = useSelected();
+  const hoveredRecordKey = useHovered();
+  /** @todo 다중 선택 기능 구현 후 코드 변경  */
+  const selectedRecordKey = [...useSelected()][0] as UIRecordKey | undefined;
   const getItemReference = useItemReference();
 
-  const { setStatus, setHovered } = useDispatcher();
+  const { setHovered, setStatus } = useDispatcher();
+
+  const { startResize, resize, endResize } = useResizeFunctions(selectedRecordKey);
+  const { startRotate, rotate, endRotate } = useRotateFunctions(selectedRecordKey);
 
   const onDocumentMouseDown = useEvent((event: MouseEvent) => {
     if (!(event.target instanceof Element && event.target.closest(`[${UIDesignToolElementDataAttributeName.id}]`))) {
@@ -43,7 +50,7 @@ export const InteractionController: React.FC<InteractionControllerProps> = React
     const handle = event.target.closest<HTMLElement>(`[${UIInteractionElementDataAttributeName.handleType}]`);
 
     /** @todo 데이터로 판단하는 대신 `fromMouse`처럼 overlay를 반환하는 메서드를 추가해 선택 영역을 클릭하면 true를 지정하도록 변경 (overlay가 바닥을 덮도록 함) */
-    const isSelectionGrabbing = isUIRecordKey(hovered);
+    const isSelectionGrabbing = isUIRecordKey(hoveredRecordKey);
     const isHandleGrabbing = handle != null;
 
     const shouldTransform = isSelectionGrabbing || isHandleGrabbing;
@@ -52,27 +59,62 @@ export const InteractionController: React.FC<InteractionControllerProps> = React
     if (shouldSelection) {
       setStatus({ statusType: StatusType.selection });
       uiControllerAPI.select([]);
+
+      // return startSelection(event);
       return;
     }
 
     if (shouldTransform) {
-      const handleType =
+      const transformMethod =
         (handle?.dataset[UIInteractionElementDataset.handleType] as Exclude<TransformMethod, 'none'> | undefined) ?? TransformMethod.move;
-      setStatus({ statusType: StatusType.transform, transformMethod: handleType });
-      /** @todo 다중 선택 기능 구현 후 조건 추가  */
-      uiControllerAPI.select(isUIRecordKey(hovered) ? [hovered] : []);
+      setStatus({ statusType: StatusType.transform, transformMethod });
+      if (!isHandleGrabbing) {
+        /** @todo 다중 선택 기능 구현 후 조건 추가  */
+        uiControllerAPI.select(isUIRecordKey(hoveredRecordKey) ? [hoveredRecordKey] : []);
+      }
+
+      // if (transformMethod === TransformMethod.move) {
+      //   return startMove(event);
+      // }
+
+      if (transformMethod === TransformMethod.resize) {
+        return startResize(event);
+      }
+
+      if (transformMethod === TransformMethod.rotate) {
+        return startRotate(event);
+      }
+
       return;
     }
 
     uiControllerAPI.select([]);
   });
 
-  const onDocumentMouseUp = useEvent(() => {
+  const onDocumentMouseUp = useEvent((event: MouseEvent) => {
     setStatus({ statusType: StatusType.idle });
+
+    // if (status.statusType === StatusType.selection) {
+    //   return endSelection(event);
+    // }
+
+    if (status.statusType === StatusType.transform) {
+      // if (status.transformMethod === TransformMethod.move) {
+      //   return endMove(event);
+      // }
+
+      if (status.transformMethod === TransformMethod.resize) {
+        return endResize(event);
+      }
+
+      if (status.transformMethod === TransformMethod.rotate) {
+        return endRotate(event);
+      }
+    }
   });
 
-  const onDocumentMouseMove = useEvent(() => {
-    console.log({ ...status, hovered, selected: [...selected] });
+  const onDocumentMouseMove = useEvent((event: MouseEvent) => {
+    console.log({ ...status, hovered: hoveredRecordKey, selected: selectedRecordKey });
 
     if (status.statusType === StatusType.idle) {
       const target = uiSelectorAPI.fromMouse();
@@ -85,9 +127,25 @@ export const InteractionController: React.FC<InteractionControllerProps> = React
       return;
     }
 
-    // if (status.statusType === StatusType.selection) {}
+    // if (status.statusType === StatusType.selection) {
+    //   return selection(event);
+    // }
 
-    // if (status.statusType === StatusType.transform) {}
+    if (status.statusType === StatusType.transform) {
+      // if (status.transformMethod === TransformMethod.move) {
+      //   return move(event);
+      // }
+
+      if (status.transformMethod === TransformMethod.resize) {
+        return resize(event);
+      }
+
+      if (status.transformMethod === TransformMethod.rotate) {
+        return rotate(event);
+      }
+
+      return;
+    }
   });
 
   useEventListener(document, 'mousedown', onDocumentMouseDown);
