+++ packages/ui-design-tool/components/InteractionController/useRotateFunctions.tsx
@@ -1,0 +1,140 @@
+import useBrowserMeta from '@/hooks/useBrowserMeta';
+import useDispatcher from '@/hooks/useDispatcher';
+import useUIController from '@/hooks/useUIController';
+import useUISelector from '@/hooks/useUISelector';
+import { UIRecordRect, UIRecordRectInit } from '@/types/Geometry';
+import { UIRecordKey } from '@/types/Identifier';
+import { isUIRecordKey } from '@/utils/model';
+import { getComputedUIRecordStyleValue } from '@/utils/style';
+import { cursor } from '@pigyuma/design-system/extensions';
+import { setRef } from '@pigyuma/react-utils';
+import { calcDegreesBetweenCoords, isEqual, pick, toDegrees360 } from '@pigyuma/utils';
+import { useCallback, useRef } from 'react';
+import { useItemReference } from '../UIDesignToolProvider/UIDesignToolProvider.context';
+
+const getTransformedRect = (rect: UIRecordRect, mousePoint: { x: number; y: number }, handleCoordDegrees: number) => {
+  const newRectInit: UIRecordRectInit = pick(rect, ['x', 'y', 'width', 'height', 'rotate']);
+
+  let rotate = calcDegreesBetweenCoords(
+    {
+      x: rect.x + rect.width / 2,
+      y: rect.y + rect.height / 2,
+    },
+    mousePoint,
+  );
+  rotate -= handleCoordDegrees;
+  rotate += rect.rotate;
+  rotate = Math.round(rotate);
+
+  newRectInit.rotate = toDegrees360(rotate);
+
+  return UIRecordRect.fromRect(newRectInit);
+};
+
+const getRotateCursor = (element: Element, mousePoint: { x: number; y: number }) => {
+  const rect = element.getBoundingClientRect();
+  return cursor.rotatePoint({ x: rect.x + rect.width / 2, y: rect.y + rect.height / 2 }, mousePoint);
+};
+
+export default function useRotateFunctions(recordKey: UIRecordKey | undefined) {
+  const transformInitialRectRef = useRef<UIRecordRect>();
+  const transformLastRectRef = useRef<UIRecordRect>();
+
+  const rotateHandleCoordDegreesRef = useRef<number>();
+
+  const uiControllerAPI = useUIController();
+  const uiSelectorAPI = useUISelector();
+
+  const getBrowserMeta = useBrowserMeta();
+  const getItemReference = useItemReference();
+
+  const { setCursor } = useDispatcher();
+
+  const startRotate = useCallback(
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    (event: MouseEvent) => {
+      const record = isUIRecordKey(recordKey) ? getItemReference(recordKey) : undefined;
+      const target = isUIRecordKey(recordKey) ? uiSelectorAPI.query({ key: recordKey }) : undefined;
+      if (record == null || target == null) {
+        return;
+      }
+
+      const browserMeta = getBrowserMeta();
+      const mouseMeta = browserMeta.mouse;
+      const mouseOffsetPoint = { x: mouseMeta.offsetX, y: mouseMeta.offsetY };
+      const mouseClientPoint = { x: mouseMeta.clientX, y: mouseMeta.clientY };
+
+      const rotate = parseFloat(getComputedUIRecordStyleValue(target, 'rotate')) || 0;
+      const rect = UIRecordRect.fromRect({ ...UIRecordRect.fromElement(target).toJSON(), rotate });
+
+      setRef(transformInitialRectRef, rect);
+      setRef(transformLastRectRef, transformInitialRectRef.current);
+      setRef(
+        rotateHandleCoordDegreesRef,
+        calcDegreesBetweenCoords({ x: rect.x + rect.width / 2, y: rect.y + rect.height / 2 }, mouseOffsetPoint),
+      );
+      setCursor(getRotateCursor(target, mouseClientPoint));
+    },
+    [recordKey, uiSelectorAPI, getBrowserMeta, getItemReference, setCursor],
+  );
+
+  const endRotate = useCallback(
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    (event: MouseEvent) => {
+      const record = isUIRecordKey(recordKey) ? getItemReference(recordKey) : undefined;
+      if (record == null) {
+        return console.error(`UIRecord '${recordKey}' not found.`);
+      }
+
+      const target = isUIRecordKey(recordKey) ? uiSelectorAPI.query({ key: recordKey }) : undefined;
+      if (target == null) {
+        return console.error(`Element with recordKey of '${recordKey}' not found.`);
+      }
+
+      const rect = transformLastRectRef.current ?? UIRecordRect.fromElement(target);
+
+      setRef(transformInitialRectRef, undefined);
+      setRef(transformLastRectRef, undefined);
+      setRef(rotateHandleCoordDegreesRef, undefined);
+      uiControllerAPI.setRect(record.key, rect);
+    },
+    [recordKey, uiControllerAPI, uiSelectorAPI, getItemReference],
+  );
+
+  const rotate = useCallback(
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    (event: MouseEvent) => {
+      const record = isUIRecordKey(recordKey) ? getItemReference(recordKey) : undefined;
+      if (record == null) {
+        return console.error(`UIRecord '${recordKey}' not found.`);
+      }
+
+      const target = isUIRecordKey(recordKey) ? uiSelectorAPI.query({ key: recordKey }) : undefined;
+      if (target == null) {
+        return console.error(`Element with recordKey of '${recordKey}' not found.`);
+      }
+
+      const initialRect = transformInitialRectRef.current;
+      if (initialRect == null) {
+        throw new Error("'rotate' event was not properly initialized.");
+      }
+
+      const browserMeta = getBrowserMeta();
+      const mouseMeta = browserMeta.mouse;
+      const mouseOffsetPoint = { x: mouseMeta.offsetX, y: mouseMeta.offsetY };
+      const mouseClientPoint = { x: mouseMeta.clientX, y: mouseMeta.clientY };
+      const handleCoordDegrees = rotateHandleCoordDegreesRef.current ?? 0;
+
+      const newRect = getTransformedRect(initialRect, mouseOffsetPoint, handleCoordDegrees);
+
+      if (!isEqual(newRect.toJSON(), transformLastRectRef.current?.toJSON())) {
+        setRef(transformLastRectRef, newRect);
+        uiControllerAPI.setRect(record.key, newRect);
+      }
+      setCursor(getRotateCursor(target, mouseClientPoint));
+    },
+    [recordKey, uiControllerAPI, uiSelectorAPI, getBrowserMeta, getItemReference, setCursor],
+  );
+
+  return { startRotate, endRotate, rotate };
+}
